+++ modules/firefox.nix
@@ -1,4 +1,35 @@
 {
+  darwinModule = { pkgs, lib, ... }: {
+    system.activationScripts.applications.text = lib.mkAfter ''
+      install -o root -g wheel -m0555 -d "/Applications/Firefox.app"
+
+      rsyncFlags=(
+        # mtime is standardized in the nix store, which would leave only file size to distinguish files.
+        # Thus we need checksums, despite the speed penalty.
+        --checksum
+        # Converts all symlinks pointing outside of the copied tree (thus unsafe) into real files and directories.
+        # This neatly converts all the symlinks pointing to application bundles in the nix store into
+        # real directories, without breaking any relative symlinks inside of application bundles.
+        # This is good enough, because the make-symlinks-relative.sh setup hook converts all $out internal
+        # symlinks to relative ones.
+        --copy-unsafe-links
+        --archive
+        --delete
+        --chmod=-w
+        --no-group
+        --no-owner
+      )
+
+      ${lib.getExe pkgs.rsync} "''${rsyncFlags[@]}" \
+        ${pkgs.firefox-bin-unwrapped}/Applications/Firefox.app/ /Applications/Firefox.app
+    '';
+
+    system.defaults.CustomUserPreferences."org.mozilla.firefox".EnterprisePoliciesEnabled =
+      true;
+    system.defaults.CustomUserPreferences."org.mozilla.firefox".DisableAppUpdate =
+      true;
+  };
+
   homeModule = { config, pkgs, lib, ... }:
     let
       inherit (pkgs.stdenv) hostPlatform;
@@ -16,7 +47,7 @@
 
       programs.firefox.enable = lib.mkDefault true;
       programs.firefox.package =
-        if hostPlatform.isDarwin then pkgs.firefox-bin else pkgs.firefox;
+        if hostPlatform.isDarwin then null else pkgs.firefox;
       programs.firefox.profiles.base = {
         id = 0;