+++ overlays/firefox.nix
@@ -1,0 +1,4 @@
+self: super: {
+  firefox-bin-unwrapped = super.firefox-bin-unwrapped.overrideAttrs
+    (old: { dontFixup = assert !old ? dontFixup; true; });
+}